// src/entities/Chat/model/types/Message.ts
// Generated by https://quicktype.io

export interface Message {
	id: string;
	type: number;
	content: string;
	channel_id: string;
	author: Author;
	attachments: Attachment[];
	embeds: Embed[];
	mentions: Author[];
	mention_roles: string[];
	pinned: boolean;
	mention_everyone: boolean;
	tts: boolean;
	timestamp: string;
	edited_timestamp: null;
	flags: number;
	components: unknown[];
}
// Generated by https://quicktype.io

export interface Embed {
	type:                 string;
	url:                  string;
	thumbnail:            Thumbnail;
	content_scan_version: number;
}

export interface Thumbnail {
	url:       string;
	proxy_url: string;
	width:     number;
	height:    number;
}

export interface Attachment {
	id: string;
	filename: string;
	size: number;
	url: string;
	proxy_url: string;
	content_type: string;
}

export interface Author {
	id: string;
	username: string;
	avatar: string;
	discriminator: string;
	public_flags: number;
	premium_type: number;
	flags: number;
	banner: null;
	accent_color: null;
	global_name: string;
	avatar_decoration_data: null;
	banner_color: null;
}